requirements.txt
# MSMD-Synthesizer Requirements
# Core dependencies for Phase 1: Foundational Framework

# Numerical computing
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0

# Graph analysis and manipulation
networkx>=2.6.0

# JSON handling (built-in with Python 3.6+)
# No additional package needed for JSON

# Optional: Enhanced visualization
seaborn>=0.11.0

# Optional: Progress bars for long operations
tqdm>=4.62.0

# Development and testing (optional)
pytest>=6.2.0
pytest-cov>=2.12.0

# Phase 4: AI retrieval (Gemini AI)
google-generativeai>=0.3.0
python-dotenv>=0.19.0
# Keep scikit-learn as optional fallback
# scikit-learn>=1.0.0

# Phase 5: Benchmark reporting
tabulate>=0.9.0